backend/requirements.txt
fastapi
uvicorn
opencv-python
ultralytics
pymongo
python-multipart
python-dotenv